mcp_servers/sql_execution/pyproject.toml
[tool.pytest.ini_options]
testpaths = ["mcp_servers/sql_execution_server/tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark test as async",
]

[project]
name = "mercor-rls-sql-execution"
version = "0.1.0"
description = "Mercor RLS SQL Execution MCP"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aioboto3>=15.4.0",
    "aiobotocore>=2.25.0",
    "asyncer>=0.0.9",
    "asyncpg>=0.30.0",
    "datadog-api-client>=2.44.0",
    "fastmcp>=2.12.4,<2.14.0",
    "httpx>=0.27.0",
    "litellm>=1.77.7",
    "loguru>=0.7.3",
    "mcp-schema",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.6",
    "ruff>=0.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportUnusedCallResult = "none"
failOnWarnings = false
exclude = ["scripts", "packages", "modal_labs.py", "modal_helpers.py", ".venv", ".git", ".tmp", "venv", "mcp_servers", "smoke_test", "tests"]

[tool.ruff.lint]
# Enable linting rules
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # Line too long
]

[tool.coverage.run]
source = ["mcp_servers/sql_execution_server/tools"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]


[tool.uv.sources]
mcp-schema = { path = "packages/mcp_schema" }
